move rgbToHex to colors.js
